feature(middleware): add middleware for routes

+++ routes/web.php
@@ -24,24 +24,24 @@ Route::get('/quotes/{quote}', [QuoteController::class, 'show']);
 
 Route::get('/movies/{movie:slug}', [MovieController::class, 'index']);
 
-Route::get('/login', [SessionController::class, 'create']);
+Route::get('/login', [SessionController::class, 'create'])->middleware('guest');
 Route::post('/login', [SessionController::class, 'store']);
 Route::post('/logout', [SessionController::class, 'destroy'])->middleware('auth');
 
-Route::get('/admin', [AdminController::class, 'index']);
+Route::get('/admin', [AdminController::class, 'index'])->middleware('auth');
 
-Route::get('admin/movies', [AdminMovieController::class, 'index']);
-Route::get('/admin/movies/create', [AdminMovieController::class, 'create']);
-Route::post('/admin/movies', [AdminMovieController::class, 'store']);
+Route::get('admin/movies', [AdminMovieController::class, 'index'])->middleware('auth');
+Route::get('/admin/movies/create', [AdminMovieController::class, 'create'])->middleware('auth');
+Route::post('/admin/movies', [AdminMovieController::class, 'store'])->middleware('auth');
 
-Route::get('admin/movies/{movie}/edit', [AdminMovieController::class, 'edit']);
-Route::patch('admin/movies/{movie}', [AdminMovieController::class, 'update']);
-Route::delete('admin/movies/{movie}', [AdminMovieController::class, 'destroy']);
+Route::get('admin/movies/{movie}/edit', [AdminMovieController::class, 'edit'])->middleware('auth');
+Route::patch('admin/movies/{movie}', [AdminMovieController::class, 'update'])->middleware('auth');
+Route::delete('admin/movies/{movie}', [AdminMovieController::class, 'destroy'])->middleware('auth');
 
-Route::get('admin/quotes', [AdminQuoteController::class, 'index']);
-Route::get('admin/quotes/create', [AdminQuoteController::class, 'create']);
-Route::post('admin/quotes', [AdminQuoteController::class, 'store']);
+Route::get('admin/quotes', [AdminQuoteController::class, 'index'])->middleware('auth');
+Route::get('admin/quotes/create', [AdminQuoteController::class, 'create'])->middleware('auth');
+Route::post('admin/quotes', [AdminQuoteController::class, 'store'])->middleware('auth');
 
-Route::get('admin/quotes/{quote}/edit', [AdminQuoteController::class, 'edit']);
-Route::patch('admin/quotes/{quote}', [AdminQuoteController::class, 'update']);
-Route::delete('admin/quotes/{quote}', [AdminQuoteController::class, 'destroy']);
+Route::get('admin/quotes/{quote}/edit', [AdminQuoteController::class, 'edit'])->middleware('auth');
+Route::patch('admin/quotes/{quote}', [AdminQuoteController::class, 'update'])->middleware('auth');
+Route::delete('admin/quotes/{quote}', [AdminQuoteController::class, 'destroy'])->middleware('auth');
